+++ cypress/integration/2.3-multiProduct.spec.js
@@ -42,6 +42,7 @@ describe('Multi Product Testcase', () => {
 
   it('Verify tax and total', updateRetry(3), () => {
     // Verify Tax
+    cy.qe(`Verifying the tax should have ${tax}`)
     cy.get(selectors.TaxAmtLabel).last().should('have.text', tax)
     // Verify Total
     cy.verifyTotal(totalAmount)