+++ cypress/support/commands.js
@@ -3,6 +3,12 @@ import { FAIL_ON_STATUS_CODE, VTEX_AUTH_HEADER } from './common/constants.js'
 // Order Tax API Test Case
 Cypress.Commands.add('orderTaxApi', (requestPayload, tax) => {
   cy.getVtexItems().then(vtex => {
+    cy.qe(`
+    curl --location --request POST '${vtex.baseUrl}/cybersource/checkout/order-tax' \
+--header 'X-VTEX-API-AppKey: AppKey' \
+--header 'X-VTEX-API-AppToken: AppToken' \
+--data-raw 'tData":{req.body}'
+    `)
     cy.request({
       method: 'POST',
       url: `${vtex.baseUrl}/${
@@ -28,6 +34,7 @@ Cypress.Commands.add('orderTaxApi', (requestPayload, tax) => {
 })
 
 Cypress.Commands.add('checkForTaxErrors', () => {
+  cy.qe(`communication error with Tax System has occurred should not exist`)
   cy.contains('communication error with Tax System has occurred', {
     timeout: 2000,
   }).should('not.exist')